+++ micro-ui/web/micro-ui-internals/packages/libraries/src/services/elements/CustomService.js
@@ -1,12 +1,12 @@
 import { Request } from "../atoms/Utils/Request";
 
 export const CustomService = {
-  getResponse: ({ url, params, body, plainAccessRequest,useCache=true,userService=true,setTimeParam=true,userDownload=false,auth=true, headers={}}) =>  Request({
+  getResponse: ({ url, params, body, plainAccessRequest,useCache=true,userService=true,setTimeParam=true,userDownload=false,auth=true, headers={}, method="POST"}) =>  Request({
       url: url,
       data: body,
       useCache,
       userService,
-      method: "POST",
+      method: method,
       auth:auth,
       params: params,
       headers: headers,